requirements.txt
twilio==8.12.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.36
nasdaq-data-link==1.0.4 